refactor columns, clenaup
